refactor(ios): Move iOS viewport scaling to a dedicated provider
