+++ www/app/Chat.tsx
@@ -103,37 +103,6 @@ interface HonchoResponse {
   content: string;
 }
 
-function useIOSViewportScale() {
-  useEffect(() => {
-    // Only run on iOS devices
-    const isIOS = /iPad|iPhone|iPod/.test(navigator.userAgent);
-
-    if (isIOS) {
-      let viewport = document.querySelector(
-        'meta[name="viewport"]'
-      ) as HTMLMetaElement;
-
-      // Create one if it doesn't exist
-      if (!viewport) {
-        viewport = document.createElement('meta');
-        viewport.name = 'viewport';
-        document.head.appendChild(viewport);
-      }
-
-      // Set the viewport content
-      viewport.content =
-        'width=device-width, initial-scale=1.0, maximum-scale=1.0';
-
-      // Cleanup function
-      return () => {
-        if (viewport) {
-          viewport.content = 'width=device-width, initial-scale=1.0';
-        }
-      };
-    }
-  }, []);
-}
-
 export default function Chat({
   initialUserId,
   initialEmail,
@@ -165,8 +134,6 @@ export default function Chat({
   const input = useRef<ElementRef<'textarea'>>(null);
   const messageContainerRef = useRef<ElementRef<'section'>>(null);
   useAutoScroll(messageContainerRef);
-
-  useIOSViewportScale();
 
   const messageListRef = useRef<MessageListRef>(null);
   const firstChat = useMemo(() => {
